+++ runtime/src/commonMain/kotlin/dev/zacsweers/metro/internal/SetFactory.kt
@@ -16,6 +16,8 @@
 package dev.zacsweers.metro.internal
 
 import dev.zacsweers.metro.Provider
+import kotlin.js.JsStatic
+import kotlin.jvm.JvmStatic
 
 /**
  * A [Factory] implementation used to implement [Set] bindings. This factory always returns a new
@@ -81,6 +83,8 @@ private constructor(
   public companion object {
     private val EMPTY_FACTORY: Factory<Set<Any>> = InstanceFactory(emptySet())
 
+    @JvmStatic
+    @JsStatic
     public fun <T> empty(): Factory<Set<T>> {
       @Suppress("UNCHECKED_CAST")
       return EMPTY_FACTORY as Factory<Set<T>>
@@ -92,6 +96,8 @@ private constructor(
      *
      * Equivalent to [builder] with one individual provider but skips the [Builder] allocation.
      */
+    @JvmStatic
+    @JsStatic
     public fun <T : Any> singleton(provider: Provider<T>): Factory<Set<T>> =
       SingletonSetFactory(provider)
 
@@ -99,6 +105,8 @@ private constructor(
      * Constructs a new [Builder] for a [SetFactory] with `individualProviderSize` individual
      * `Provider<T>` and `collectionProviderSize` `Provider<Collection<T>>` instances.
      */
+    @JvmStatic
+    @JsStatic
     public fun <T : Any> builder(
       individualProviderSize: Int,
       collectionProviderSize: Int,